[#1196] Move datapusher code into separate extension
