+++ ckan/config/deployment.ini_tmpl
@@ -79,6 +79,7 @@ ckan.site_id = default
 ## Plugins Settings
 
 # Note: Add ``datastore`` to enable the CKAN DataStore
+#       Add ``datapusher`` to enable DataPusher
 #       Add ``pdf_preview`` to enable the resource preview for PDFs
 #		Add ``resource_proxy`` to enable resorce proxying and get around the
 #		same origin policy
