+++ ckanext/datapusher/logic/auth.py
@@ -1,0 +1,9 @@
+import ckanext.datastore.logic.auth as auth
+
+
+def datapusher_submit(context, data_dict):
+    return auth.datastore_auth(context, data_dict)
+
+
+def datapusher_status(context, data_dict):
+    return auth.datastore_auth(context, data_dict)
