+++ ckanext/datastore/logic/auth.py
@@ -1,7 +1,7 @@
 import ckan.plugins as p
 
 
-def _datastore_auth(context, data_dict, privilege='resource_update'):
+def datastore_auth(context, data_dict, privilege='resource_update'):
     if not 'id' in data_dict:
         data_dict['id'] = data_dict.get('resource_id')
     user = context.get('user')
@@ -19,28 +19,20 @@ def _datastore_auth(context, data_dict, privilege='resource_update'):
 
 
 def datastore_create(context, data_dict):
-    return _datastore_auth(context, data_dict)
+    return datastore_auth(context, data_dict)
 
 
 def datastore_upsert(context, data_dict):
-    return _datastore_auth(context, data_dict)
+    return datastore_auth(context, data_dict)
 
 
 def datastore_delete(context, data_dict):
-    return _datastore_auth(context, data_dict)
+    return datastore_auth(context, data_dict)
 
 
 def datastore_search(context, data_dict):
-    return _datastore_auth(context, data_dict, 'resource_show')
-
-
-def datapusher_submit(context, data_dict):
-    return _datastore_auth(context, data_dict)
-
-
-def datapusher_status(context, data_dict):
-    return _datastore_auth(context, data_dict)
+    return datastore_auth(context, data_dict, 'resource_show')
 
 
 def datastore_change_permissions(context, data_dict):
-    return _datastore_auth(context, data_dict)
+    return datastore_auth(context, data_dict)
